Add yet more Javadoc and restore the independence of the `NamespaceSupplier`.
